pyproject.toml
[project]
name = "creativethinkingaiagent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.21.0",
]
